Update re-root.py
Some changes to make it work with current versions of Radarr.
NetImport was changed to ImportLists in the Radarr database, and the default name for the database is now "radarr.db"
Removed some unnecessary comments.
Converted docstrings to
the triple double-quoted string form.
Changed printout to show what string it is changing, AND what it is changing to.